requirements.txt
python-dotenv
Flask==2.3.3
Flask-Bcrypt
Flask-Cors
flask-restx
Pillow
pandas
requests
huggingface_hub
langchain
fpdf
pandas
langchain
pypdf
docarray
huggingface_hub
langchain-community
panel
sentence_transformers
langchain-huggingface
text-generation
transformers
google-search-results
numexpr
langchainhub
sentencepiece
jinja2
bitsandbytes
accelerate
pandas-datareader
pdfplumber